+++ src/object/defineFrozen.js
@@ -1,7 +1,0 @@
-'use strict';
-
-import define from './define.js';
-
-export default function defineFrozen (props = {}, obj = {}) {
-    return Object.freeze(define(props, obj));
-}