+++ src/object/defineSealed.js
@@ -1,7 +1,0 @@
-'use strict';
-
-import define from './define.js';
-
-export default function defineSealed (props = {}, obj = {}) {
-    return Object.seal(define(props, obj));
-}