+++ src/regexp/is.js
@@ -1,5 +1,7 @@
 'use strict';
 
+export const PROTO_RGX = '[object RegExp]';
+
 export default function isRegExp (val) {
-    return Object.prototype.toString.call(val) === '[object RegExp]';
+    return Object.prototype.toString.call(val) === PROTO_RGX;
 }